+++ packages/stream_chat/lib/src/core/models/user.g.dart
@@ -31,6 +31,7 @@ User _$UserFromJson(Map<String, dynamic> json) => User(
       teamsRole: (json['teams_role'] as Map<String, dynamic>?)?.map(
         (k, e) => MapEntry(k, e as String),
       ),
+      avgResponseTime: (json['avg_response_time'] as num?)?.toInt(),
     );
 
 Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{