+++ packages/stream_chat_persistence/test/src/mapper/user_mapper_test.dart
@@ -19,6 +19,7 @@ void main() {
       online: math.Random().nextBool(),
       banned: math.Random().nextBool(),
       teamsRole: const {'teamId': 'role', 'teamId2': 'role2'},
+      avgResponseTime: 120,
       extraData: {'test_extra_data': 'extraData'},
     );
     final user = entity.toUser();
@@ -32,6 +33,7 @@ void main() {
     expect(user.online, entity.online);
     expect(user.banned, entity.banned);
     expect(user.teamsRole, entity.teamsRole);
+    expect(user.avgResponseTime, entity.avgResponseTime);
     expect(user.extraData, entity.extraData);
   });
 
@@ -46,6 +48,7 @@ void main() {
       online: math.Random().nextBool(),
       banned: math.Random().nextBool(),
       teamsRole: const {'teamId': 'role', 'teamId2': 'role2'},
+      avgResponseTime: 120,
       extraData: const {'test_extra_data': 'extraData'},
     );
     final entity = user.toEntity();
@@ -59,6 +62,7 @@ void main() {
     expect(entity.online, user.online);
     expect(entity.banned, user.banned);
     expect(entity.teamsRole, user.teamsRole);
+    expect(entity.avgResponseTime, user.avgResponseTime);
     expect(entity.extraData, user.extraData);
   });
 }