+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -9,6 +9,10 @@
 - Fixed draft message retrieval logic where channel drafts were incorrectly attached to all messages
   instead of only thread drafts being attached to their respective parent messages.
 
+✅ Added
+
+- Added support for `User.avgResponseTime` field.
+
 ## 10.0.0-beta.3
 
 - Included the changes from version [`9.14.0`](https://pub.dev/packages/stream_chat_persistence/changelog).